integrate more dep.

+++ arrow-datafusion/arrow-rs/arrow-array/src/array/list_array.rs
@@ -35,7 +35,7 @@ pub trait OffsetSizeTrait: ArrowNativeType + std::ops::AddAssign + Integer {
     const PREFIX: &'static str;
 }
 
-impl OffsetSizeTrait for i8 {
+impl OffsetSizeTrait for i16 {
     const IS_LARGE: bool = false;
     const PREFIX: &'static str = "";
 }
